+++ src/main/java/ca/corefacility/bioinformatics/irida/web/controller/api/projects/RESTProjectSamplesController.java
@@ -243,6 +243,7 @@ public class RESTProjectSamplesController {
 	 * @param seqeuncerId the string id of the sample
 	 * @return The found sample
 	 */
+	@Deprecated()
 	@RequestMapping(value = "/api/projects/{projectId}/samples/bySequencerId/{seqeuncerId}", method = RequestMethod.GET)
 	public ModelAndView getProjectSampleBySequencerId(@PathVariable Long projectId, @PathVariable String seqeuncerId) {
 		Project p = projectService.read(projectId);
@@ -251,6 +252,28 @@ public class RESTProjectSamplesController {
 
 		Link withSelfRel = linkTo(
 				methodOn(RESTProjectSamplesController.class).getSample(sampleBySampleId.getId())).withSelfRel();
+		String href = withSelfRel.getHref();
+
+		RedirectView redirectView = new RedirectView(href);
+
+		return new ModelAndView(redirectView);
+	}
+
+	/**
+	 * Get samples by a given string name
+	 *
+	 * @param projectId   the Project to get samples from
+	 * @param sampleName  String containing sample name
+	 * @return The found sample
+	 */
+	@RequestMapping(value = "/api/projects/{projectId}/samples/bySampleName", method = RequestMethod.GET)
+	public ModelAndView getProjectSampleBySampleName(@PathVariable Long projectId, @RequestParam String sampleName) {
+		Project p = projectService.read(projectId);
+
+		Sample sampleBySampleName = sampleService.getSampleBySampleName(p, sampleName);
+
+		Link withSelfRel = linkTo(
+				methodOn(RESTProjectSamplesController.class).getSample(sampleBySampleName.getId())).withSelfRel();
 		String href = withSelfRel.getHref();
 
 		RedirectView redirectView = new RedirectView(href);